lingbin-temp-update: temp-commit-at-20240106
lingbin-temp-update: temp-commit-at-20240204

lingbin-temp-update: temp-commit-at-20240215

lingbin-temp-update: temp-commit-at-20240215-2

lingbin-temp-update: temp-commit-at-20240216

lingbin-temp-update: temp-commit-at-20240217

lingbin-temp-update: temp-commit-at-20240224

lingbin-temp-update: temp-commit-at-20240226

lingbin-temp-update: temp-commit-at-20240308

lingbin-temp-update: temp-commit-at-20240316

lingbin-temp-update: temp-commit-at-20240331

lingbin-temp-update: temp-commit-at-20240417

lingbin-temp-update: temp-commit-at-20240420

lingbin-temp-update: temp-commit-at-20240426

lingbin-temp-update: temp-commit-at-20240514

lingbin-temp-update: temp-commit-at-20240516

lingbin-temp-update: temp-commit-at-20240623

lingbin-temp-update: temp-commit-at-20240627

lingbin-temp: fix Driver.h

lingbin-temp: fix VectorHasher.cpp for 20240514

lingbin-temp: fix VectorHasher.cpp for 20240630

lingbin-temp: fix VectorHasher.cpp for 20240702

Fix PeriodicStatsReporter.h on MocOS(clang)

lingbin-temp: temp-commit-at-20240712

lingbin-temp: temp-commit-at-20240714: MmapArea

lingbin-temp: temp-commit-at-20240714: MemoryAllocator

lingbin-temp: temp-commit-at-20240714: MemoryAllocator

lingbin-temp: temp-commit-at-20240717: Driver

lingbin-temp: temp-commit-at-20240720: Task

lingbin-temp: temp-commit-at-20240806: exec-ExchangeClient

lingbin-temp: temp-commit-at-20240806: type-Filter

lingbin-temp: temp-commit-at-20240806: connector

lingbin-temp: temp-commit-at-20240810: connector

lingbin-temp: temp-commit-at-20240823: Vector

lingbin-temp: temp-commit-at-20240825: Connector

lingbin-temp: temp-commit-at-20240826: HashProbe-remove copy

+++ velox/common/base/RawVector.h
@@ -47,7 +47,7 @@ class raw_vector {
     }
   }
 
-  // Constructs  a copy of 'other'. See operator=. 'data_' must be copied.
+  // Constructs a copy of 'other'. See operator=. 'data_' must be copied.
   raw_vector(const raw_vector<T>& other) {
     *this = other;
   }
@@ -56,8 +56,6 @@ class raw_vector {
     *this = std::move(other);
   }
 
-  // Moves 'other' to this, leaves 'other' empty, as after default
-  // construction.
   void operator=(const raw_vector<T>& other) {
     resize(other.size());
     if (other.data_) {
@@ -68,6 +66,7 @@ class raw_vector {
     }
   }
 
+  // Moves 'other' to this, leaves 'other' empty, as after default construction.
   void operator=(raw_vector<T>&& other) noexcept {
     data_ = other.data_;
     size_ = other.size_;
@@ -159,7 +158,7 @@ class raw_vector {
  private:
   // Adds 'bytes' to the address 'pointer'.
   inline T* addBytes(T* pointer, int32_t bytes) {
-    return reinterpret_cast<T*>(reinterpret_cast<uint64_t>(pointer) + bytes);
+    return reinterpret_cast<T*>(reinterpret_cast<uintptr_t>(pointer) + bytes);
   }
 
   // Size with one full width SIMD load worth data above and below, rounded to
@@ -200,12 +199,12 @@ class raw_vector {
   int32_t capacity_{0};
 };
 
-// Returns a pointer to 'size' int32_t's with consecutive values
-// starting at 0. There are at least kPadding / sizeof(int32_t) values
-// past 'size', so that it is safe to access the returned pointer at maximum
-// SIMD width. Typically returns preallocated memory but if this is
-// not large enough,resizes and initializes 'storage' to the requested
-// size and returns storage.data().
+// Returns a pointer to 'size' int32_t's with consecutive values starting at 0.
+// There are at least kPadding / sizeof(int32_t) values past 'size', so that it
+// is safe to access the returned pointer at maximum SIMD width. Typically,
+// returns pre-allocated memory but if this is not large enough, otherwise,
+// resizes and initializes 'storage' to the requested size and returns
+// storage.data().
 const int32_t*
 iota(int32_t size, raw_vector<int32_t>& storage, int32_t offset = 0);
